Library

Public API

Internal API

Composites.compliance — Method.

compliance(A::Array{<:Real,2}, B::Array{<:Real,2}, D::Array{<:Real,2})

Returns alpha, beta, and delta

source

Composites.getABD! — Method.

`getABD(matid::AbstractArray{<:Integer,1},
    nply::AbstractArray{<:Integer,1}, tply::AbstractArray{<:Real,1},
    theta::AbstractArray{<:Real,1},
    q::AbstractArray{<:AbstractArray{<:Real,2},1})`

`getABD(lam::Laminate, q::AbstractArray{<:AbstractArray{<:Real,2},1})`

Returns A, B, and D matrices

Arguments:

matid::AbstractArray{<:Integer,1}: material id of each lamina

nply::AbstractArray{<:Integer,1}: number of plies in each lamina

tply::AbstractArray{<:Real,1}: thickness of a ply (m) in each lamina

theta::AbstractArray{<:Real,1}: orientation (deg) of each lamina

q::AbstractArray{<:AbstractArray{<:Real,2}}: Stiffness matrix of each lamina

source

Composites.getQ — Function.

`getQ(e1::Real, e2::Real, g12::Real, nu12::Real, theta::Real=0.0)`

`getQ(mat::Material, theta::Real)`

`getQ(mat::Material, lam::Laminate)`

Returns Q matrix rotated theta degrees

source

Composites.getmatfail — Method.

`getmatfail(sigma1::Real, sigma2::Real, tau12::Real, xt::Real,
    xc::Real, yt::Real, yc::Real, s::Real, method::String)`

`getmatfail(sigma1::Real, sigma2::Real, tau12::Real, mat::Material,
    method::String)`

`getmatfail(stress::AbstractArray{<:Real,1}, mat::Material, method::String)`

`getmatfail(stress::AbstractArray{<:Real,1},xt::Real,xc::Real,yt::Real,
    yc::Real,s::Real, method::String)`

`getmatfail(stress::AbstractArray{<:AbstractArray{<:Real,1},1},
    mat::AbstractArray{Material}, lam::Laminate, method::String)`

Determines ply failure according to specified theory. Values greater than one signify ply failure. Choose between "maxstress", "tsaiwu", and "hashinrotem". Safety factors are stored in the second return argument.

Theory Specific Output

Maximum Ply Stress:

Longitudinal tension

Longitudinal compression

Transverse tension

Transverse compression

Positive shear

Negative shear

Tsai-Wu

Failure Criterion

Hashin-Rotem

Fiber failure in tension

Fiber failure in compression

Matrix failure in tension

Matrix failure in compression

source

Composites.getplystrain — Method.

`getplystrain(nply::AbstractArray{<:Integer,1}, tply::AbstractArray{<:Real,1},
    theta::AbstractArray{<:Real,1}, resultantstrain::AbstractArray{<:Real,1})`

`getplystrain(lam::Laminate, resultantstrain::AbstractArray{<:Real,1})`

Calculates strains in each ply aligned with principal material direction.

source

Composites.getplystress — Method.

`getplystress(plystrain::AbstractArray{<:AbstractArray{<:Real,1},1},
    q::AbstractArray{<:AbstractArray{<:Real,2},1},
    matid::AbstractArray{<:Integer,1})`

`getplystress(plystrain::AbstractArray{<:AbstractArray{<:Real,1},1},
    q::AbstractArray{<:AbstractArray{<:Real,2},1}, lam::Laminate)`

Calculates ply stresses from ply strains and material stiffness matrices (q)

source

Composites.getz — Method.

`getz(tply::AbstractArray{<:Real,1}, nply::AbstractArray{<:Integer,1})`

`getz(lam::Laminate)`

Returns a laminate's z-coordinates (coordinates of top and bottom of laminas) given the thickness of plies in each lamina and the number of plies in each lamina.

source

Composites.localbuckling — Method.

localbuckling(A::Array{<:Real,2}, D::Array{<:Real,2}, b::Real)

Calculates local buckling load and strains. b is the panel width.

assumes: balanced, symmetric, large aspect ratio, simply supported, uniaxial compression, flat rectangular plate. buckling strain is output as a positive value

see chapter on Structural Component Design Techniques from Alastair Johnson section 6.2: Design of composite panels

source

Composites.rotQ! — Method.

`rotQ(q::AbstractArray{<:Real,2}, theta::Real)`

`rotQ(qxx::Real, q12::Real, q22::Real, q66::Real, theta::Real)`

Rotates Q matrix by theta degrees

source

Composites.rotstrain! — Method.

`rotstrain(strain,theta)`

Rotates strain theta degrees. strain = [eps1,eps2,gamma12]

source

Composites.rotstress! — Method.

`rotstress(stress,theta)`

Rotates stress theta degrees. stress = [sigma1,sigma2,tau12]

source

Composites.Laminate — Type.

`Laminate{R<:Real}(matid, nply, tply, theta)`

Returns struct containing Laminate properties

Arguments/Fields

matid::Array{Int, 1}: Material id for each lamina

nply::Array{Int, 1}: number of plies in each lamina

tply::Array{R, 1}: ply thickness for each lamina

theta::Array{Int, 1}: orientation of each lamina (degrees)

source

Composites.Material — Type.

`Material{R<:Real}(e1,e2,g12,nu12,rho,xt,xc,yt,yc,s,t)`

Returns composite type containing composite (in-plane) material properties.

Arguments/Fields

e1: E1

e2: E2

g12: G12

nu12: poisson's Ratio

rho: density

xt: longitudinal tensile ultimate strength

xc: longitudinal compressive ultimate strength

yt: transverse tensile ultimate strength

yc: transverse compressive ultimate strength

s: shear ultimate strength

t: ply thickness

source